+++ PaymentMethods/Exceptions/CheckoutValidationException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Heidelpay\Gateway\PaymentMethods\Exceptions;
+
+
+use Exception;
+
+class CheckoutValidationException extends Exception
+{
+
+}